requirements.txt
tweepy
google-api-python-client
google-auth
google-generativeai
